requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
boto3==1.40.10
botocore==1.40.10
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-anymail==13.0.1
django-cors-headers==4.7.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
ecdsa==0.19.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
s3transfer==0.13.1
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
